refactor: extract service logic to new file
